claimguard/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
cryptography>=42.0.0
slowapi>=0.1.9
PyJWT>=2.8.0
limits>=3.13.0

# Blockchain dependencies
web3>=6.11.0
eth-account>=0.10.0
py-solc-x>=2.0.3

# IPFS/Pinata dependencies
aiohttp>=3.9.0
aiofiles>=23.2.0

# Firestore (claim storage in non-test environments)
firebase-admin>=6.0.0

# Additional utilities
requests>=2.31.0

# Document text extraction (PDF + images via Tesseract when installed)
pypdf>=4.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
networkx>=3.2.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# Multi-agent orchestration (optional LLM path via CREW_USE_LLM=1)
crewai>=0.70.0
crewai-tools>=0.12.0
# LangChain chat models — foundation for LLM routing; CrewAI unwraps these via create_llm()
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
